+++ loginCore/build.gradle
@@ -10,7 +10,7 @@ android {
     compileSdk 33
 
     defaultConfig {
-        minSdk 21
+        minSdk 24
         targetSdk 33
 
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
