+++ settings.gradle
@@ -20,3 +20,4 @@ include ':createAccountCore'
 include ':composeCore'
 include ':uiCore'
 include ':loginCompose'
+include ':createAccountCompose'
